Almost working example of incremental progress streaming.
Summary:Dune and other CLI apps use the \r to reset the cursor position, but we
don't flush until we see a newline. This change will flush before newlines too.
The only problem is that dune currently still thinks we are in a tty. That's probably
just due to our little bash wrapper. We could reimplement that in native and then
the problem should go away. Sending this diff out so someone can run with it.

Test Plan:

Reviewers:

CC: